change client_pool (#1171)
* change client_pool

* retrigger ci
